Revert special handling of tmux
This reverts the core of 9c36094 and bccd342.

I've learnt that it is not necessary to wrap the escape code for tmux,
but that screen can apply the settings per pane.

I had `set -g terminal-overrides "xterm*:XT:smcup@:rmcup@"` in
tmux.conf, but have `set -ag terminal-overrides ",xterm*:smcup@:rmcup@"` now.

Thanks to @nicm on #tmux.

Screen appears to only recognize it when wrapped specifically.